+++ src/js/lib/utils/bolt.ts
@@ -1,7 +1,7 @@
-import CSInterface from "@/lib/cep/csinterface";
-import Vulcan, { VulcanMessage } from "@/lib/cep/vulcan";
+import CSInterface from '@/lib/cep/csinterface';
+import Vulcan, { VulcanMessage } from '@/lib/cep/vulcan';
 import { ns } from '@/../shared/shared';
-import { fs } from "@/lib/cep/node";
+import { fs } from '@/lib/cep/node';
 
 export const csi = new CSInterface();
 export const vulcan = new Vulcan();
@@ -22,12 +22,12 @@ export const vulcan = new Vulcan();
 export const evalES = (script: string, isGlobal = false): Promise<string> => {
   return new Promise(function (resolve, reject) {
     const pre = isGlobal
-      ? ""
+      ? ''
       : `var host = typeof $ !== 'undefined' ? $ : window; host["${ns}"].`;
     const fullString = pre + script;
-    console.log("evalES", fullString);
+    console.debug(fullString, { module: 'extendscript' });
     csi.evalScript(
-      "try{" + fullString + "}catch(e){alert(e);}",
+      'try{' + fullString + '}catch(e){alert(e);}',
       (res: string) => {
         resolve(res);
       }
@@ -35,7 +35,7 @@ export const evalES = (script: string, isGlobal = false): Promise<string> => {
   });
 };
 
-import type { Scripts } from "@esTypes/index";
+import type { Scripts } from '@esTypes/index';
 
 type ArgTypes<F extends Function> = F extends (...args: infer A) => any
   ? A
@@ -80,7 +80,7 @@ export const evalTS = <
         console.log(JSON.stringify(arg));
         return `${JSON.stringify(arg)}`;
       })
-      .join(",");
+      .join(',');
     csi.evalScript(
       `try{
           var host = typeof $ !== 'undefined' ? $ : window;
@@ -92,10 +92,11 @@ export const evalTS = <
         }`,
       (res: string) => {
         try {
-          if (res === "undefined" || res === undefined) return resolve(undefined as ReturnType<Func>);
+          if (res === 'undefined' || res === undefined)
+            return resolve(undefined as ReturnType<Func>);
           const parsed = JSON.parse(res);
-          if (parsed.name === "ReferenceError") {
-            console.error("REFERENCE ERROR");
+          if (parsed.name === 'ReferenceError') {
+            console.error('REFERENCE ERROR');
             reject(parsed);
           } else {
             resolve(parsed);
@@ -123,7 +124,7 @@ export const evalFile = (file: string) => {
 
 export const initBolt = (log = true) => {
   if (window.cep) {
-    const extRoot = csi.getSystemPath("extension");
+    const extRoot = csi.getSystemPath('extension');
     const jsxSrc = `${extRoot}/jsx/index.js`;
     const jsxBinSrc = `${extRoot}/jsx/index.jsxbin`;
     if (fs.existsSync(jsxSrc)) {
@@ -136,7 +137,7 @@ export const initBolt = (log = true) => {
   }
 };
 
-export const posix = (str: string) => str.replace(/\\/g, "/");
+export const posix = (str: string) => str.replace(/\\/g, '/');
 
 export const openLinkInBrowser = (url: string) => {
   if (window.cep) {
@@ -163,7 +164,7 @@ export const getAppBackgroundColor = () => {
 export const subscribeBackgroundColor = (callback: (color: string) => void) => {
   const getColor = () => {
     const newColor = getAppBackgroundColor();
-    console.log("BG Color Updated: ", { rgb: newColor.rgb });
+    console.log('BG Color Updated: ', { rgb: newColor.rgb });
     const { r, g, b } = newColor.rgb;
     return `rgb(${r}, ${g}, ${b})`;
   };
@@ -171,7 +172,7 @@ export const subscribeBackgroundColor = (callback: (color: string) => void) => {
   callback(getColor());
   // listen for changes
   csi.addEventListener(
-    "com.adobe.csxs.events.ThemeColorChanged",
+    'com.adobe.csxs.events.ThemeColorChanged',
     () => callback(getColor()),
     {}
   );
@@ -230,7 +231,7 @@ export const selectFolder = (
     dir
   ) as IOpenDialogResult;
   if (result.data?.length > 0) {
-    const folder = decodeURIComponent(result.data[0].replace("file://", ""));
+    const folder = decodeURIComponent(result.data[0].replace('file://', ''));
     callback(folder);
   }
 };
@@ -247,7 +248,7 @@ export const selectFile = (
     dir
   ) as IOpenDialogResult;
   if (result.data?.length > 0) {
-    const folder = decodeURIComponent(result.data[0].replace("file://", ""));
+    const folder = decodeURIComponent(result.data[0].replace('file://', ''));
     callback(folder);
   }
 };